alx_travel_app/alx_travel_app/requirement.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
djangorestframework==3.16.1
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
pika==1.3.2
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
